# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Fretes",
    "version": "1.0.0",
    "description": "Documentação da API de Fretes"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "paths": {
    "/createFreight": {
      "post": {
        "summary": "Cria um novo frete",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Freight"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Frete cadastrado com sucesso!"
          },
          "400": {
            "description": "Erro: Frete não cadastrado com sucesso!"
          }
        }
      }
    },
...
    "/createProduct": {
      "post": {
        "summary": "Cria um novo produto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto cadastrado com sucesso!"
          },
          "400": {
            "description": "Erro: Produto não cadastrado com sucesso!"
          }
        }
      }
    },
...
    "/createUser": {
      "post": {
        "summary": "Cria um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário cadastrado com sucesso!"
          },
          "400": {
            "description": "Erro: Usuário não cadastrado com sucesso!"
          }
        }
      }
    },
...
    "/createVehicle": {
      "post": {
        "summary": "Cria um novo veículo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vehicle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Veículo cadastrado com sucesso!"
          },
          "400": {
            "description": "Erro: Veículo não cadastrado com sucesso!"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Freight": {
        "type": "object",
        "properties": {
          "product": { "type": "string" },
          "product_wheight": { "type": "string" },
          "vehicle": { "type": "string" },
          "vehicle_wheight": { "type": "string" },
          "freight": { "type": "number" },
          "delivery_value": { "type": "number" },
          "kms": { "type": "string" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "product_name": { "type": "string" },
          "product_type": { "type": "string" },
          "product_wheight": { "type": "string" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "user_name": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "vehicle_name": { "type": "string" },
          "vehicle_type": { "type": "string" },
          "max_wheight": { "type": "string" }
        }
      }
    }
  }
}